pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.5"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
ipython = "^8.24.0"
django-autoslug = "^1.9.9"
django-crispy-forms = "^2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
